+++ battlesim/distrib/__init__.py
@@ -1,0 +1,2 @@
+from ._distrib2 import Sampling
+from ._composite import Composite