+++ tests/test_battle.py
@@ -16,11 +16,6 @@ sys.path.insert(0, "../")
 import battlesim as bsm
 
 
-def test_define_battle():
-    # battle object requires input file
-    b = bsm.Battle("../datasets/starwars-clonewars.csv")
-
-
 def test_battle_attributes():
     # battle object requires input file
     b = bsm.Battle("../datasets/starwars-clonewars.csv")
@@ -31,8 +26,6 @@ def test_battle_attributes():
 
     with pytest.raises(AttributeError):
         b.composition_
-    with pytest.raises(AttributeError):
-        b.n_allegiance_
 
     return b
 
@@ -63,108 +56,6 @@ def test_battle_create_army():
     # created normally.
     b.create_army([("B1 battledroid", 10), ("Clone Trooper", 10)])
     return b
-
-
-def test_apply_position():
-    # battle object requires input file
-    b = bsm.Battle("../datasets/starwars-clonewars.csv")
-
-    # first, attempt to call without create_army
-    with pytest.raises(AttributeError):
-        b.apply_position("gaussian")
-
-    # created normally.
-    b.create_army([("B1 battledroid", 100), ("Clone Trooper", 100)])
-
-    # try parameter diffs
-    with pytest.raises(TypeError):
-        b.apply_position()
-    with pytest.raises(TypeError):
-        b.apply_position(1)
-
-    # apply_pos with false distribution
-    with pytest.raises(ValueError):
-        b.apply_position("fake")
-
-    b.apply_position("exponential")
-    # check distribution using KS-test > 0.05
-    assert kstest(b.M_["x"], "expon")[1] > 0.05, "P-value for KS-test not > 0.05, fail"
-    assert kstest(b.M_["y"], "expon")[1] > 0.05, "P-value for KS-test not > 0.05, fail"
-    b.apply_position("laplace")
-    assert kstest(b.M_["x"], "laplace")[1] > 0.05, "P-value for KS-test not > 0.05, fail"
-
-    #
-    with pytest.raises(ValueError):
-        b.apply_position(["fake"])
-        b.apply_position([("Hello")])
-
-    # dist for each unit - works
-    b.apply_position(["gaussian", "uniform"])
-    assert kstest(b.M_["x"][:100], "norm")[1] > 0.01, "P-value for KS-test not > 0.05, fail"
-    assert kstest(b.M_["y"][-100:], "uniform")[1] > 0.01, "P-value for KS-test not > 0.05, fail"
-
-    # we don't test bsm.Distribution objects here
-    b.apply_position({"name": "gaussian", "loc": 0., "scale": 1.})
-
-    # ignore wrong keyword
-    b.apply_position({"n": "gaussian", "loc": 0., "scale": 1.})
-    with pytest.raises(ValueError):
-        b.apply_position({"name": 10, "loc": 0., "scale": 1.})
-    with pytest.raises(ValueError):
-        b.apply_position({"name": str(10), "loc": 0., "scale": 1.})
-    with pytest.raises(TypeError):
-        # wrong loc, scale parameters
-        b.apply_position({"name": "gaussian", "loc": "hello", "scale": 1.})
-
-
-def test_set_initial_ai():
-    # battle object requires input file
-    b = bsm.Battle("../datasets/starwars-clonewars.csv")
-
-    # set AI before definition
-    with pytest.raises(AttributeError):
-        b.set_initial_ai(["nearest", "pack"])
-    with pytest.raises(AttributeError):
-        b.set_initial_ai("nearest")
-
-    # created normally.
-    b.create_army([("B1 battledroid", 100), ("Clone Trooper", 100)])
-
-    # call error
-    with pytest.raises(ValueError):
-        b.set_initial_ai("hello")
-    with pytest.raises(ValueError):
-        b.set_initial_ai(["nearest", "hello"])
-
-    # call normally
-    b.set_initial_ai(["nearest", "random"])
-    # check composition
-    assert set(b.composition_["init_ai"].values) == set(["random", "nearest"])
-
-
-def test_set_rolling_ai():
-    # battle object requires input file
-    b = bsm.Battle("../datasets/starwars-clonewars.csv")
-
-    # set AI before definition
-    with pytest.raises(AttributeError):
-        b.set_rolling_ai(["nearest", "random"])
-    with pytest.raises(AttributeError):
-        b.set_rolling_ai("nearest")
-
-    # created normally.
-    b.create_army([("B1 battledroid", 100), ("Clone Trooper", 100)])
-
-    # call error
-    with pytest.raises(ValueError):
-        b.set_rolling_ai("hello")
-    with pytest.raises(ValueError):
-        b.set_rolling_ai(["nearest", "hello"])
-
-    # call normally
-    b.set_rolling_ai(["nearest", "random"])
-    # check composition
-    assert set(b.composition_["rolling_ai"].values) == set(["random", "nearest"])
 
 
 def test_simulate():